+++ src/main/java/com/scg/stop/domain/file/domain/File.java
@@ -4,7 +4,7 @@ import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
-import com.scg.stop.domain.gallery.domain.Gallery;
+import com.scg.stop.gallery.domain.Gallery;
 import com.scg.stop.notice.domain.Notice;
 import com.scg.stop.event.domain.EventNotice;
 import com.scg.stop.global.domain.BaseTimeEntity;
