+++ src/main/java/com/scg/stop/gallery/dto/request/GalleryRequest.java
@@ -1,4 +1,4 @@
-package com.scg.stop.domain.gallery.dto.request;
+package com.scg.stop.gallery.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
