+++ src/main/java/com/scg/stop/gallery/repository/GalleryRepository.java
@@ -1,14 +1,12 @@
-package com.scg.stop.domain.gallery.repository;
+package com.scg.stop.gallery.repository;
 
-import com.scg.stop.domain.gallery.domain.Gallery;
+import com.scg.stop.gallery.domain.Gallery;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface GalleryRepository extends JpaRepository<Gallery, Long> {
